feat(api): Support enabling fiat at runtime
This PR will always register the `FiatAuthenticationFilter` in the
authentication chain _but_ allow for it to be disabled at runtime
depending on the dynamic config service.

The `FiatAuthenticationFilter` has been modified to no-op when
fiat would have previously been disabled.

A metric is also emitted (`fiat.enabled` to track whether fiat is
enabled on a service/instance basis).