feat(core): Support a fallback to legacy account permissions
Should `fiat` ever be unavailable, this supports a fallback to a
`UserPermission` that is constructed based on the contents of
the `AuthenticatedRequest`.

In the event of a legacy fallback, access to _all_ applications will be
explicitly granted.

It is not on by default and must be explicited opt'd enabled via:

```
fiat:
  services:
    legacyFallback: true
```

or by setting the `fiat.legacyFallback.enabled` dynamic property.