fix(roles): Avoid filtering when `allowAccessToUnknownApplications` is enabled
This flag was originally added to allow sites with large numbers of
applications to avoid needlessly serializing the same set of applications
for each user.

Unfortunately it doesn't play nice when an application was explicitly
filtered out because a user was not granted access to it!

This PR will no longer filter applications but rather include them in
the permission with _no_ authorizations.

It only applies when `allowAccessToUnknownApplications` is enabled.

```
[
  {
    "name": "clouddriver",
    "authorizations": [
      "READ",
      "WRITE"
    ]
  },
  {
    "name": "orca",
    "authorizations": [

    ]
  }
]
```

vs

```
[
  {
    "name": "clouddriver",
    "authorizations": [
      "READ",
      "WRITE"
    ]
  }
]
```