Lock redis db when syncing
The sync endpoint is modifying the redis roles database. The problem is
that there may be a concurrent operation that also modifies the
database, making it prone to race conditions.

This patch just locks the redis database when modifying the roles from
the period operation as well as from the sync endpoint.