feat(roles): Add execute authz when its missing.
When an application does not explicity specify which groups have
`EXECUTE` permission, add `EXECUTE` permissions to groups that have
`READ` or `WRITE` based on the value of `fiat.executeFallback`. For
backwards compatibility, `fiat.executeFallback` defaults to `READ`
i.e. groups that have `READ` on an application will now have `EXECUTE`
as well.