refactor(fiat-roles): add a ResourcePermissionProvider interface
There are two of these that are required to be provided for Fiat to
startup correctly: one for Accounts and one for Applications. There is
a default `@Configuration` that maintains the status quo by providing
default implementations that simply read the permissions directly off of
the objects, but alternate configurations can be provided that retrieve
permissions in other ways.