fix(api): fix serialization issues with new Authorization enum values.
This reverts a well intended change where the resulting Permissions object always had an entry for each
Authorization type. This change makes rolling out new values in this enum painful to coordinate across
services.

This reverts the behaviour to now only store an explicit entry in the Map if there are specifically roles
defined for that Authorization type