fix: Application name mismatch when fiat permissions coming from different sources. #5156
This is part of: spinnaker/spinnaker#5156.

Fixed the name mismatch bug in ResourcePrefixPermissionSource.java

Resource(Application) always returns the name in CAPITAL letters.

Permission prefixes added in external sources in fiat-local.yml has the application name in lowercase letters.

Modified the code to check these names with

1)prefix.equalsIgnoreCase(resource.getName());

2) prefixWithoutStar = prefixWithoutStar.toUpperCase();

Modified the ResourcePrefixPermissionSourceSpec.java to fix the testcases.