fix(java11): fix Stream#count() use under Java 11 (#656)
Cherry-picking part of #630

UserRoleSyncer was using stream.map().count() to run an operation on
each element and then return the number of elements in the stream. But
this doesn't work in Java 11 because "An implementation may choose to
not execute the stream pipeline (either sequentially or in parallel) if
it is capable of computing the count directly from the stream source."
The Java 11 docs (unlike the Java 8 docs) include an explicit example of
this exact thing we're doing and why it doesn't work.

Fixes spinnaker/spinanker#5466
Fixes spinnaker/spinnaker#5687