perf(auth/ldap): Limit search object size
Currently when fiat sync LDAP roles with the new Group -> User mapping
the response is huge, as mentioned in the original commit. To workaround
this issue this change adds the wanted attributes to the search, making
the search only return those attributes.

To be able to use "attributes to return" feature we also have to set a
search scope. Using the default scope of SUBTREE_SCOPE caused errors on
my test setup, and changing to OBJECT_SCOPE solved those issues. Unit tests
however still pass in both cases.

The delimiter for the GroupUserAttributes is configurable, but will
default to space, as it's the default for ldapsearch.