feat(perf): cache the unrestricted users permissions
The unrestricted user is loaded from redis each time a lookup for a user happens, and
this user may end up with a lot of data associated with it if there are a large number
of secured entities set up for anonymous read access.

This saves the step of fetching and deserializing that data if there is a copy cached.

Adds a new last modified key to track (via redis server time) the last time the
unrestricted user was written to redis, and uses that to ensure a reload happens when
the permissions are rebuilt.