fix(repo): don't swallow exceptions in PermissionRepository.get
Allows RedisPermissionRepository.get to fail with an exception
instead of masking failure with an Optional.empty

This will allow better fallback behaviour and retries instead
of just appearing as an unknown user.

Adds retries for reading user data from redis to handle pool
exhaustion or redis timeouts and introduces an overall
timeout for the user read in case to prevent excessive retries
from introducing unacceptable latency.

The RedisPermissionRepository.get by default is set to
time out after 1 second but this value is configurable with
the property 'fiat.redis.repository.get-permission-timeout'

There is also a timeout for the call to determine whether the
UNRESTRICTED user has changed. By default this is 50 milliseconds
but is configurable with the property
'fiat.redis.repository.check-last-modified-timeout'