Unify fetching of user permissions
This change shares the read from Redis code between `get()`, `getAllById()`, and `getAllByRoles()`.

Impact to performance should be minimal as the unrestrictied user permissions cache is now used in the `getAllBy*` cases. There is however an additional `SISMEMBER` call per-user fetched rather than a single `SSCAN` and test on the Java side.