feat(sql): Add an SQL permission repository
This PR maintains backwards compatibility and does not modify existing
controllers.

Implements a permission repository backed by SQL. No migration is
provided as the data is recreated on the next role sync or at startup.

Enabling SQL as the permission repository

```yaml
sql:
  enabled: true
  connectionPools:
    default:
      jdbcUrl: jdbc:mysql://.....
      user: ...
      password: ...
      default: true
  migration:
      jdbcUrl: jdbc:mysql://.....
      user: ...
      password: ...

permissionRepository:
  redis:
    enabled: false
  sql:
    enabled: true
```

Both MySQL and Postgres are supported.

Integration tests have been updated to enable testing with the SQL
provider (using testcontainers), but still default to Redis only
as that's faster.