fix(config): Remove unsupported config keys
This change removes the configuration key `endpoints.health.sensitive`
from `fiat.yml` as it's no longer supported.

Currently when starting Fiat the following message is printed:
```
The use of configuration keys that are no longer supported was found in the environment:

Property source 'applicationConfig: [classpath:/fiat.yml]':
	Key: endpoints.health.sensitive
		Line: 13
		Reason: Endpoint sensitive flag is no longer customizable as Spring Boot no longer provides a customizable security auto-configuration .
```

With this change that message is no longer printed.