fix: improve performance of role sync with SQL
The permission rebuild process during role synchronization only requires
the set of roles that each user has. This change modifies the SQL and
Redis `PermissionRepository` instances to only return the roles.

Note that only the SQL repository makes use of this change to improve
performance. The Redis one is currently unoptimised.