perf(fiat-roles): Cache long-running roles' synchronization process
With a large number of concurrently requested roles synchronizations, application uses a lot of memory and runs out of available sql connections in connection pool. We are able to assign long-running processes to Future<>, store it in in-memory cache and utilize it instead of starting new synchronization in every thread.

Change is backwards compatible, and it is not enabled by default. Property fiat.role.sync-cache.enabled governs whether to use old implementation or new one.